Fix divergence_diagnostic.py to use fp64 matching test_correctness.py
Changed both output casts in divergence_diagnostic.py from np.float32 to
np.float64 (lines 66 and 94), matching test_correctness.py's fp64 contract.
Regenerated divergence_results.txt under consistent fp64 precision.

+++ benchmark/divergence_diagnostic.py
@@ -63,7 +63,7 @@ import sys, numpy as np
 from {mod} import run
 s, y, o = run({call})
 s(); y()
-a = o.to_numpy().flatten().astype(np.float32)
+a = o.to_numpy().flatten().astype(np.float64)
 np.save("{npy}", a)
 nan_ct = int(np.isnan(a).sum())
 print(f"n={{len(a)}} nan={{nan_ct}}")
@@ -91,7 +91,7 @@ import sys, numpy as np
 sys.path.insert(0, "{BD}")
 from {mod} import *
 _, _, o = {call}
-a = np.array(o).flatten().astype(np.float32) if not isinstance(o, np.ndarray) else o.flatten().astype(np.float32)
+a = np.array(o).flatten().astype(np.float64) if not isinstance(o, np.ndarray) else o.flatten().astype(np.float64)
 np.save("{npy}", a)
 nan_ct = int(np.isnan(a).sum())
 print(f"n={{len(a)}} nan={{nan_ct}}")
